terraform/functions/get_translations/requirements.txt
google-cloud-firestore>=2.11.0
